First attempt at power.cpp split
